Change the input r to long int

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -9,7 +9,7 @@
  * Return: value of the last digit
  */
 
-int print_last_digit(int r)
+int print_last_digit(long int r)
 {
 	int i = 0;
